requirements.txt
# Core AI Framework
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.20
google-generativeai>=0.8.0

# Backend API
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.0
sse-starlette==1.8.2

# Frontend UI
streamlit==1.31.0
streamlit-agraph==0.0.45

# Database & Data Handling
neo4j==5.16.0
pandas==2.2.0
beautifulsoup4==4.12.3
requests==2.31.0

# Async HTTP for dynamic scraping
aiohttp==3.9.1

# Environment Variables
python-dotenv==1.0.0

# Additional utilities
pyvis==0.3.2
markdown==3.5.1
